requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
djangorestframework==3.15.2
flake8==7.1.2
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
pycodestyle==2.12.1
pyflakes==3.2.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
